requirements.txt
manim
python-dotenv
